(FMT): move block classes to separate package
