+++ src/main/kotlin/org/rust/ide/formatter/blocks/RustMacroArgFmtBlock.kt
@@ -1,14 +1,15 @@
-package org.rust.ide.formatter
+package org.rust.ide.formatter.blocks
 
 import com.intellij.formatting.*
 import com.intellij.lang.ASTNode
+import org.rust.ide.formatter.RustFmtContext
 
 class RustMacroArgFmtBlock(
     node: ASTNode,
     alignment: Alignment?,
     indent: Indent?,
     wrap: Wrap?,
-    ctx: RustFmtBlockContext
+    ctx: RustFmtContext
 ) : AbstractRustFmtBlock(node, alignment, indent, wrap, ctx) {
 
     override fun getSpacing(child1: Block?, child2: Block): Spacing? = Spacing.getReadOnlySpacing()
